+++ src/listener/builder.rs
@@ -22,6 +22,8 @@ where
     #[cfg(feature = "async-ports")]
     handle: Option<Handle>,
     tick_interval: Option<Duration>,
+    #[cfg(feature = "async-ports")]
+    async_tick: bool,
     poll_timeout: Duration,
 }
 
@@ -38,6 +40,8 @@ where
             handle: None,
             poll_timeout: Duration::from_millis(10),
             tick_interval: None,
+            #[cfg(feature = "async-ports")]
+            async_tick: false,
         }
     }
 }
@@ -51,25 +55,37 @@ where
     /// # Errors
     ///
     /// - if there are async ports defined, but no handle is set.
-    pub(crate) fn start(self) -> Result<EventListener<U>, ListenerError> {
+    #[allow(unused_mut)] // mutability is necessary when "async-ports" is active
+    pub(crate) fn start(mut self) -> Result<EventListener<U>, ListenerError> {
+        #[cfg(feature = "async-ports")]
+        let start_async =
+            !self.async_ports.is_empty() || (self.async_tick && self.tick_interval.is_some());
         #[cfg(not(feature = "async-ports"))]
         let store_tx = false;
         #[cfg(feature = "async-ports")]
-        let store_tx = !self.async_ports.is_empty();
-        #[allow(unused_mut)] // mutability is necessary when "async-ports" is active
-        let mut res = EventListener::new(self.poll_timeout).start(
-            self.sync_ports,
-            self.tick_interval,
-            store_tx,
-        );
+        let store_tx = start_async;
+        #[cfg(not(feature = "async-ports"))]
+        let sync_tick_interval = self.tick_interval;
+        #[cfg(feature = "async-ports")]
+        let sync_tick_interval = self.tick_interval.take_if(|_| !self.async_tick);
+        let mut res = EventListener::new(self.poll_timeout);
+
+        // dont start a sync worker if there are no sync tasks
+        if !self.sync_ports.is_empty() || sync_tick_interval.is_some() {
+            res = res.start(self.sync_ports, sync_tick_interval, store_tx);
+        }
 
         // dont start a taskpool without any actual tasks
         #[cfg(feature = "async-ports")]
-        if !self.async_ports.is_empty() {
+        if start_async {
             let Some(handle) = self.handle else {
                 return Err(ListenerError::NoHandle);
             };
-            res = res.start_async(self.async_ports, handle);
+            res = res.start_async(
+                self.async_ports,
+                handle,
+                self.tick_interval.take_if(|_| self.async_tick),
+            );
         }
 
         Ok(res)
@@ -178,6 +194,14 @@ where
         self.handle = Some(handle);
         self
     }
+
+    /// Change the way [`Event::Tick`](crate::Event::Tick) is emitted from being on a [`SyncPort`] to be a [`AsyncPort`].
+    #[cfg(feature = "async-ports")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "async-ports")))]
+    pub fn async_tick(mut self, value: bool) -> Self {
+        self.async_tick = value;
+        self
+    }
 }
 
 #[cfg(test)]
@@ -262,5 +286,30 @@ mod test {
         let builder = builder.async_port(port);
 
         assert_eq!(builder.start().unwrap_err(), ListenerError::NoHandle);
+    }
+
+    #[tokio::test]
+    #[cfg(feature = "async-ports")]
+    async fn should_spawn_async_ticker() {
+        use tokio::time::sleep;
+
+        use crate::Event;
+
+        let builder = EventListenerCfg::<MockEvent>::default()
+            .with_handle(Handle::current())
+            .async_tick(true)
+            .tick_interval(Duration::from_millis(10));
+        assert!(builder.async_ports.is_empty());
+        assert_eq!(Handle::current().metrics().num_alive_tasks(), 0);
+
+        let mut listener = builder.start().unwrap();
+        assert_eq!(Handle::current().metrics().num_alive_tasks(), 1);
+        assert!(listener.thread.is_none()); // there are no sync ports or tasks, so no sync worker
+        assert!(listener.taskpool.is_some());
+        sleep(Duration::from_millis(25)).await; // wait for at least 2 events
+
+        listener.stop().unwrap();
+        assert_eq!(listener.poll(), Ok(Some(Event::Tick)));
+        assert_eq!(listener.poll(), Ok(Some(Event::Tick)));
     }
 }